ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "gradio>=5.49.1",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
